requirements.txt
greenlet==0.4.17
isort==5.7.0
msgpack==1.0.2
neovim==0.3.1
PyLTSpice==1.1
pynvim==0.4.2
PyYAML==5.3.1
